+++ base/src/org/compiere/model/I_M_PriceList.java
@@ -238,12 +238,12 @@ public interface I_M_PriceList
 	/** Set Price Precision.
 	  * Precision (number of decimals) for the Price
 	  */
-	public void setPricePrecision (BigDecimal PricePrecision);
+	public void setPricePrecision (int PricePrecision);
 
 	/** Get Price Precision.
 	  * Precision (number of decimals) for the Price
 	  */
-	public BigDecimal getPricePrecision();
+	public int getPricePrecision();
 
     /** Column name Updated */
     public static final String COLUMNNAME_Updated = "Updated";
